requirements.txt
amqp==5.3.1
asgiref==3.11.0
attrs==25.4.0
billiard==4.2.4
celery==5.6.2
celery-types==0.24.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.6
Django==5.2.9
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-environ==0.12.0
django-filter==25.2
django-redis==6.0.0
django-stubs==5.2.8
django-stubs-ext==5.2.8
django-timezone-field==7.2.1
djangorestframework==3.16.1
djangorestframework-stubs==3.16.6
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
idna==3.11
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
packaging==25.0
pillow==12.1.0
prompt_toolkit==3.0.52
psycopg==3.3.2
psycopg-binary==3.3.2
PyJWT==2.10.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
redis==7.1.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
sqlparse==0.5.5
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20250913
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.2
vine==5.1.0
wcwidth==0.2.14
